Dev (#11)
* Mejora en el backend de la Api Catalogo para subir imagenes, ademas de arreglo responsivo en una vista admin

* Ultimas Mejoras

* Feature/arreglo de los modals (#9)

* feat(admin,catalogo): dividir modal (genérico/imagenes), fix hooks y merge de stock desde inventario

- Crea dos modals independientes:
  - AdminModalImagenes: mantiene subida/preview y selección de imagen principal.
  - AdminModal: modal genérico sin imágenes para Usuarios/Inventario/Pedidos.
- Corrige posicionamiento y UX del modal (overlay, centrado, scroll interno).
- Soluciona warning ESLint: mueve useEffect por encima del return condicional.
- Actualiza Catalogo.jsx para calcular stock en frontend:
  - merge de urnas + inventario (suma por urnaId, considerando estado Disponible).
  - render de stock en las cards (evita "Stock undefined").

Ajustes menores:
- Prevención de duplicación de sección “Imágenes” en Urnas.

* me falto agregar los cambios de adminmodalimagenes (antiguo adminmodal)

* quite cerrar sesion redundante ekisde

* Feature/pedidos sincronizados (#10)

* se guarda la direccion del usuario al hacer la compra, aparte de eso. Al finalizar la compra la vista de pedidos se actualiza para mostrar el nuevo pedido y el stock se actualiza.

* feat(pedidos): DTOs y servicios consistentes, subtotal calculado por BD, controllers limpios

---------

+++ Back-end/pedidos-api-descansosdelrecuerdospa/src/main/java/com/descansos_del_recuerdo_spa/pedidos/controllers/DetallePedidoController.java
@@ -15,9 +15,8 @@ public class DetallePedidoController {
 
     private final DetallePedidoService detallePedidoService;
 
-    // 🔵 Obtener todos los detalles de un pedido
     @GetMapping("/pedido/{pedidoId}")
-    public ResponseEntity<List<DetallePedidoDTO>> listarDetallesPorPedido(@PathVariable Integer pedidoId) {
+    public ResponseEntity<List<DetallePedidoDTO>> listarDetallesPorPedido(@PathVariable Long pedidoId) {
         return ResponseEntity.ok(detallePedidoService.listarDetallesPorPedido(pedidoId));
     }
 }
