+++ Back-end/pedidos-api-descansosdelrecuerdospa/src/main/java/com/descansos_del_recuerdo_spa/pedidos/services/DetallePedidoService.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface DetallePedidoService {
 
-    List<DetallePedidoDTO> listarDetallesPorPedido(Integer pedidoId);
+    List<DetallePedidoDTO> listarDetallesPorPedido(Long pedidoId);
 }
